+++ src/main/java/lotto/utils/Parser.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.utils;
 
 import java.util.Arrays;
 import java.util.List;
